feat: Polar-aware kcal goal, macro split fixes, UI improvements
- Polar mode: use BMR + tracked activity kcal (no PAL multiplier) as
  daily kcal goal; GetKcalGoalUsecase and HomeBloc now consistent
- HomeBloc syncs calorieGoal to TrackedDayEntity on each load so
  diary and dashboard always show the same goal
- MacroSplitDialog: restored g/kg protein/fat mode for Polar users;
  fixed .toInt() truncation causing macro split mismatch; added
  CalendarDayBloc + DiaryBloc refresh on save
- LowKcalWarningCard: added dismiss button with SharedPreferences
  persistence (shared_preferences dependency added)
- Profile page: show Polar/InfluxDB weight with '(Polar)' label
- intake_vertical_list: + button fixed at index 0 so it stays visible
  regardless of how many items are in the meal list
- Fixed stray merge conflict markers in config_entity.dart
- Fixed HomeBloc locator registration (16th locator() call)
- Fixed duplicate named args in HomeBloc HomeLoadedState constructor
- Fixed AddIntakeUsecase test constructor (PolarInfluxdbDataSource arg)

+++ lib/features/settings/presentation/bloc/settings_bloc.dart
@@ -203,9 +203,9 @@ class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
     double fatGoalPct,
   ) async {
     await _addConfigUsecase.setConfigMacroGoalPct(
-      carbGoalPct.toInt() / 100,
-      proteinGoalPct.toInt() / 100,
-      fatGoalPct.toInt() / 100,
+      carbGoalPct / 100,
+      proteinGoalPct / 100,
+      fatGoalPct / 100,
     );
   }
